🛡️ Guardian (JULES): Consolidate date formatting utilities
WHAT:
- Merged src/lib/date.ts into src/lib/format.ts
- Updated 8 files with new import paths
- Deleted obsolete date.ts file and its test file, merging tests into format.test.ts

WHY:
- Date utilities were a separate module causing architectural smell while date formatting naturally belongs in format.ts as canonical location.

JULES CONTEXT:
- Checked .Jules/task-log.md: No Autonomous conflicts
- Session logged: .Jules/guardian/2026-08-06/
- Reference: .Jules/JULES.md

IMPACT:
- Files: 2 → 1 (-50%)
- Import paths: standardized to @/lib/format

VERIFICATION:
- All tests pass
- TypeScript check: ✓
- Linting: ✓
- JULES compliance: ✓
