+++ .Jules/guardian/2026-08-06/session-report.md
@@ -1,0 +1,7 @@
+## 2026-08-06 - Consolidated date formatting utilities
+
+**Target:** src/lib/date.ts, src/lib/format.ts
+**Learning:** Date and time utilities must be centralized in format.ts instead of date.ts.
+**Action:** Merged date.ts functions into format.ts and deleted date.ts.
+**JULES Check:** Verified no autonomous tasks in task-log.md.
+**Conflicts Avoided:** None.
