+++ .Jules/task-log.md
@@ -1,0 +1,12 @@
+# Task Log
+
+## 2026-08-06 16:34 - [GUARDIAN] Session started
+
+- Directory: .Jules/guardian/2026-08-06
+- Phase: PRE-FLIGHT
+- JULES Check: COMPLETE
+
+## 2026-08-06 - [GUARDIAN] Date consolidation complete
+
+- Linting: ✓
+- Formatting: ✓
